refactor: rename atuin-config to atuin-dotfiles
