+++ atuin-dotfiles/src/shell.rs
