+++ atuin-dotfiles/src/shell/bash.rs
