+++ atuin-dotfiles/src/shell/fish.rs
